+++ croud/__main__.py
@@ -34,6 +34,7 @@ from croud.cmd import (
     consumer_eventhub_lease_storage_connection_string_arg,
     consumer_eventhub_lease_storage_container_arg,
     consumer_schema_arg,
+    consumer_set_id_arg,
     consumer_table_arg,
     crate_password_arg,
     crate_username_arg,
@@ -56,7 +57,7 @@ from croud.cmd import (
     user_id_or_email_arg,
 )
 from croud.config import Configuration, config_get, config_set
-from croud.consumersets.commands import consumer_sets_list
+from croud.consumersets.commands import consumer_sets_edit, consumer_sets_list
 from croud.login import login
 from croud.logout import logout
 from croud.me import me
@@ -137,7 +138,39 @@ command_tree = {
                     )
                 ],
                 "calls": consumer_sets_list,
-            }
+            },
+            "edit": {
+                "help": "Edit the specified consumer set "
+                "for a CrateDB Cloud Azure IoT product",
+                "extra_args": [
+                    consumer_set_id_arg,
+                    lambda req_opt_group, opt_opt_group:
+                    consumer_eventhub_connection_string_arg(
+                        req_opt_group, opt_opt_group, False
+                    ),
+                    lambda req_opt_group, opt_opt_group:
+                    consumer_eventhub_consumer_group_arg(
+                        req_opt_group, opt_opt_group, False
+                    ),
+                    lambda req_opt_group, opt_opt_group:
+                    consumer_eventhub_lease_storage_connection_string_arg(
+                        req_opt_group, opt_opt_group, False
+                    ),
+                    lambda req_opt_group, opt_opt_group:
+                    consumer_eventhub_lease_storage_container_arg(
+                        req_opt_group, opt_opt_group, False
+                    ),
+                    lambda req_opt_group, opt_opt_group:
+                    consumer_schema_arg(
+                        req_opt_group, opt_opt_group, False
+                    ),
+                    lambda req_opt_group, opt_opt_group:
+                    consumer_table_arg(
+                        req_opt_group, opt_opt_group, False
+                    ),
+                ],
+                "calls": consumer_sets_edit,
+            },
         },
     },
     "projects": {