+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changes for croud
 Unreleased
 ==========
 
+- Added ``consumer-sets edit`` command that allows users to edit their consumer
+  sets.
+
 - Added ``consumer-sets list`` command that allows users to list their consumer
   sets and filter them by project ID, cluster ID and product ID.